+++ tests/testthat/test-gridSearch.R
@@ -92,14 +92,15 @@ test_that("The function raises errors", {
 })
 
 # TODO: Remove with version 2.0.0
-test_that("The function warns if a raster object is used", {
-  env <- raster::stack(files)
-  expect_snapshot_warning(gridSearch(m,
-                                     hypers = h,
-                                     metric = "aicc",
-                                     test = data,
-                                     save_models = FALSE,
-                                     env = env,
-                                     interactive = FALSE,
-                                     progress = FALSE))
+test_that("The function raises an error if a raster object is used", {
+  env <- terra::rast(files)
+  class(env) <- "Raster"
+  expect_snapshot_error(gridSearch(m,
+                                   hypers = h,
+                                   metric = "aicc",
+                                   test = data,
+                                   save_models = FALSE,
+                                   env = env,
+                                   interactive = FALSE,
+                                   progress = FALSE))
 })