+++ NEWS.md
@@ -1,4 +1,4 @@
-# SDMtune 1.2.1.9000
+# SDMtune 1.3.0
 * Add function to combine cross validation models
 * Improve show method for package outputs
 * Fix CRAN note for undeclared package in Rd file